deribit-options-data-collector/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0", "poetry dynamic-versioning"]
build-backend = "poetry.core.masonry.api"

[project]
name = "deribit-options-data-collector"
version = "1.0.0"
description = "Real-time Deribit options data collector with REST/WebSocket API"
readme = "README.md"
authors = [
    {name = "Crypto Quant Team", email = "quant@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "pandas>=2.1.0",
    "pyarrow>=15.0.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
    "prometheus-client>=0.19.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "apscheduler>=3.10.0",
    "orjson>=3.9.0",
    "aiometer>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-httpx>=0.28.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "types-aiohttp>=3.9.0",
    "types-PyYAML>=6.0.0",
    "types-orjson>=3.6.0",
]

[tool.poetry.scripts]
deribit-collector = "deribit_options_collector.__main__:main"

[tool.poetry]
package-mode = "true"

[tool.poetry-data-dir]
packages = [
    {from = "src", to = "deribit_options_collector"}
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "RUF",   # Ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101", "ANN101", "ANN102", "ARG001"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "aiohttp.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "websockets.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pandas.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyarrow.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "prometheus_client.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=deribit_options_collector --cov-report=term-missing --cov-report=html"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["deribit_options_collector"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
show_missing = true
fail_under = 90

[tool.coverage.html]
directory = "htmlcov"
